irecv_send_buffer: Fix DFU file upload failure due to oversized last packet
In DFU mode the last packet contains the checksum for the transferred data
(and 12 additional bytes, total 16 bytes). If the last packet is larger than
2032 bytes we would send a packet that is larger than the packet size of 2048
bytes which would make the file upload fail.
This is especially the case for the iOS 11.0.x iBSS for the iPhone 8.
With this commit the size is checked and if it would be larger than the
packet size, we split the data portion and the checksum into two separate
packets. It might even be safe to always split the last packet into data
and checksum, but until this is proven to work with all (especially older
devices) we just split when we detect the oversized packet for now.