Add configuration option to allow compiling without USB driver
Using --with-dummy, libirecovery can be compiled in a way that it will
not require any USB driver at all. This is only useful if you just want
to query libirecovery's device database by product type or hardware model,
namely using:
    irecv_devices_get_device_by_product_type();
    irecv_devices_get_device_by_hardware_model();
All other function are either no-op or return IRECV_E_UNSUPPORTED
if the library is compiled this way.