Change irecv_send_buffer to accept an options bitfield instead of just one value
This allows to specify different options. To not break existing behavior, a
value of 1 or (1 << 0) means IRECV_SEND_OPT_DFU_NOTIFY_FINISH which is used
extensively in e.g. idevicerestore. Other options are
IRECV_SEND_OPT_DFU_FORCE_ZLP which I don't remember what it was added for,
and a new option IRECV_SEND_OPT_DFU_SMALL_PKT which needed for upload in
port DFU mode, as it won't accept packets with more than 64 bytes data and
also doesn't like a CRC attached to it.